pyproject.toml
[tool.poetry]
name = "meanroishutter"
version = "0.1.0"
description = "Script to calculate the mean of an image below a specified threshold"
authors = ["Jonathan Schaffner <jonathan.schaffner@students.fhnw.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.3"
scipy = "1.9.3"
Pillow = "^9.4.0"
pytest = "^7.3.1"
matplotlib = "^3.7.1"
tifffile = "^2023.4.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
